move torchmodel and loss to another folder
