+++ pyzoo/zoo/pipeline/api/net/torch_net.py
@@ -23,65 +23,14 @@ import torch
 from pyspark import RDD
 
 from bigdl.nn.layer import Layer
-from bigdl.util.common import JTensor
-from bigdl.nn.criterion import Criterion
 from zoo import getOrCreateSparkContext
 from zoo.common.utils import callZooFunc
 from zoo.feature.image import ImageSet
-from pyspark.serializers import CloudPickleSerializer
+from zoo.tfpark.tfnet import to_sample_rdd
 
 if sys.version >= '3':
     long = int
     unicode = str
-
-
-class TorchModel(Layer):
-    """
-    TorchModel wraps a PyTorch model as a single layer, thus the PyTorch model can be used for
-    distributed inference or training.
-    The implement of TorchModel is different from TorchNet, this TorchModel is running running
-    pytorch model in an embedding Cpython interpreter, while TorchNet transfer the pytorch model
-    to TorchScript and run with libtorch.
-    """
-
-    def __init__(self, module_bytes, weights, bigdl_type="float"):
-        weights = JTensor.from_ndarray(weights)
-        self.value = callZooFunc(
-            bigdl_type, self.jvm_class_constructor(), module_bytes, weights)
-        self.bigdl_type = bigdl_type
-
-    @staticmethod
-    def from_pytorch(model):
-        """
-        Create a TorchNet directly from PyTorch model, e.g. model in torchvision.models.
-        :param model: a PyTorch model
-        """
-        weights = []
-        for param in model.parameters():
-            weights.append(param.view(-1))
-        flatten_weight = torch.nn.utils.parameters_to_vector(weights).data.numpy()
-        bys = CloudPickleSerializer.dumps(CloudPickleSerializer, model)
-        net = TorchModel(bys, flatten_weight)
-        return net
-
-
-class TorchLoss(Criterion):
-    """
-    TorchLoss wraps a loss function for distributed inference or training.
-    This TorchLoss should be used with TorchModel.
-    """
-
-    def __init__(self, criterion_bytes, bigdl_type="float"):
-        """
-        :param bigdl_type:
-        """
-        super(TorchLoss, self).__init__(None, bigdl_type, criterion_bytes)
-
-    @staticmethod
-    def from_pytorch(criterion):
-        bys = CloudPickleSerializer.dumps(CloudPickleSerializer, criterion)
-        net = TorchLoss(bys)
-        return net
 
 
 class TorchNet(Layer):
@@ -164,7 +113,6 @@ class TorchNet(Layer):
             return ImageSet(results)
         if distributed:
             if isinstance(x, np.ndarray):
-                from zoo.tfpark.tfnet import to_sample_rdd
                 data_rdd = to_sample_rdd(x, np.zeros([x.shape[0]]), getOrCreateSparkContext())
             elif isinstance(x, RDD):
                 data_rdd = x
